Gestion d'erreur et dossier de sortie
Amélioration de la gestion des erreurs. Nom du fichier de sortie généré automatiquement. Choix du dossier possible.

+++ services/CsvDocument.py
@@ -4,7 +4,8 @@
 
 
 import csv
-from MyException import MyException
+from Error import Error
+from pathlib import Path
 
 
 # Classe permettant de créer des documents CSV d'exigences, prêts à importer dans le logiciel ProjeQtOr
@@ -13,22 +14,21 @@ class CsvDocument:
     def __init__(self: object):
         pass
 
-    def write(self, file: str, requirements):
-        try:
-            print("Creation du CSV :", file)
-            with open(file, 'w') as csvfile:
-                fieldnames = ['id', 'reference', 'name', 'idRequirementType', 'idProject', 'idProduct', 'idComponent', 'externalReference',
-                              'creationDateTime', 'idContact', 'originType', 'originId', 'idBusinessFeature', 'idUrgency', 'initialDueDate',
-                              'actualDueDate', 'description', 'idRequirement', 'idStatus', 'idResource', 'idCriticality', 'idFeasibility',
-                              'idRiskLevel', 'idPriority', 'plannedWork', 'idTargetProductVersion', 'idTargetComponentVersion', 'idMilestone',
-                              'handled', 'handledDate', 'done', 'doneDate', 'idle', 'idleDate', 'cancelled', 'result', 'locked', 'idLocker',
-                              'lockedDate', 'countTotal', 'countPlanned', 'countPassed', 'countBlocked', 'countFailed', 'countIssues']
-                writer = csv.DictWriter(csvfile, delimiter=';', fieldnames=fieldnames)
+    def write(self, file: Path, requirements):
+        if file.exists():
+            raise Error(-1, "Le fichier '" + file + "' existe déjà")
 
-                writer.writeheader()
-                for requirement in requirements:
-                    writer.writerow({'externalReference': requirement.ref,
-                                     'name': requirement.title,
-                                     'description': requirement.body + '\n\n' + requirement.test})
-        except MyException as error:
-            print(f"L'erreur suivante s'est produite : {error}")
+        with open(file, 'w') as csvfile:
+            fieldnames = ['id', 'reference', 'name', 'idRequirementType', 'idProject', 'idProduct', 'idComponent', 'externalReference',
+                          'creationDateTime', 'idContact', 'originType', 'originId', 'idBusinessFeature', 'idUrgency', 'initialDueDate',
+                          'actualDueDate', 'description', 'idRequirement', 'idStatus', 'idResource', 'idCriticality', 'idFeasibility',
+                          'idRiskLevel', 'idPriority', 'plannedWork', 'idTargetProductVersion', 'idTargetComponentVersion', 'idMilestone',
+                          'handled', 'handledDate', 'done', 'doneDate', 'idle', 'idleDate', 'cancelled', 'result', 'locked', 'idLocker',
+                          'lockedDate', 'countTotal', 'countPlanned', 'countPassed', 'countBlocked', 'countFailed', 'countIssues']
+            writer = csv.DictWriter(csvfile, delimiter=';', fieldnames=fieldnames)
+
+            writer.writeheader()
+            for requirement in requirements:
+                writer.writerow({'externalReference': requirement.ref,
+                                 'name': requirement.title,
+                                 'description': requirement.body + '\n\n' + requirement.test})
